Update Simple interest

+++ python/Simple Intrest/Main.py
@@ -1,8 +1,8 @@
 //Python3
 
-p=int(input("Enter principal value"))
-r=int(input("Enter rate"))
-t=int(input("Enter value of time"))
+p=float(input("Enter principal value"))
+r=float(input("Enter rate"))
+t=float(input("Enter value of time"))
 i=(p*r*t/100)
 total=p+i
 print (i)
